correction d'erreur sur le calcul des deltaTheo
git-svn-id: http://rdinnovation.onf.fr/svn/plugin-lvox@60 f1561d03-89af-ab4a-a97c-2bf9dc1f41eb

+++ pluginlvox/tools/lvox_computetheoriticalsthread.cpp
@@ -81,7 +81,7 @@ void LVOX_ComputeTheoriticalsThread::run()
                 _outputDeltaTheoriticalGrid->setValueAtIndex(i,-1);
             } else
             {
-                _outputDeltaTheoriticalGrid->setValueAtIndex(i, _outputDeltaTheoriticalGrid->valueAtIndex(i)/_outputTheoriticalGrid->valueAtIndex(i));
+                _outputDeltaTheoriticalGrid->setValueAtIndex(i, _outputDeltaTheoriticalGrid->valueAtIndex(i)/(float)_outputTheoriticalGrid->valueAtIndex(i));
             }
         }
         _outputDeltaTheoriticalGrid->computeMinMax();
